refactor: refactor components files structure
