+++ src/components/posts/SearchInput.tsx
